+++ src/app/orders/success/page.tsx
@@ -1,41 +1,21 @@
-// src/app/order/success/page.tsx
-"use client";
-import { useEffect, useState } from "react";
-import { useSearchParams, useRouter } from "next/navigation";
+import { Suspense } from "react";
+import SuccessClient from "./success-client";
 
-export default function SuccessPage() {
-  const sp = useSearchParams();
-  const router = useRouter();
-  const orderId = sp.get("orderId");
-  const [msg, setMsg] = useState("Finalizujemy płatność…");
+export const dynamic = "force-dynamic";
 
-  useEffect(() => {
-    let t: any;
-    // Prostą pętlą „czekamy”, aż webhook zaktualizuje zamówienie (1–2s)
-    const tick = async () => {
-      if (!orderId) return;
-      const r = await fetch(`/api/orders/${orderId}`);
-      if (!r.ok) return;
-      const o = await r.json();
-      if (o?.payment_status === "paid") {
-        setMsg("Płatność potwierdzona. Dziękujemy!");
-        t = setTimeout(() => router.push(`/order/${orderId}`), 1200);
-      } else if (o?.payment_status === "failed") {
-        setMsg("Płatność nie powiodła się. Spróbuj ponownie.");
-      } else {
-        setTimeout(tick, 1000);
-      }
-    };
-    tick();
-    return () => clearTimeout(t);
-  }, [orderId, router]);
-
+export default function Page() {
   return (
-    <div className="min-h-[60vh] grid place-items-center p-6">
-      <div className="max-w-md text-center">
-        <h1 className="text-2xl font-bold mb-2">Status płatności</h1>
-        <p>{msg}</p>
-      </div>
-    </div>
+    <Suspense
+      fallback={
+        <div className="min-h-[60vh] grid place-items-center p-6">
+          <div className="max-w-md text-center">
+            <h1 className="text-2xl font-bold mb-2">Status płatności</h1>
+            <p>Ładowanie…</p>
+          </div>
+        </div>
+      }
+    >
+      <SuccessClient />
+    </Suspense>
   );
 }
