+++ src/lib/mailer.ts
@@ -1,4 +1,3 @@
-// src/lib/mailer.ts
 import { Resend } from "resend";
 import nodemailer from "nodemailer";
 
@@ -7,9 +6,10 @@ const SMTP_HOST = process.env.SMTP_HOST || "";
 const SMTP_PORT = Number(process.env.SMTP_PORT || 465);
 const SMTP_USER = process.env.SMTP_USER || "";
 const SMTP_PASS = process.env.SMTP_PASS || "";
+const EMAIL_FROM = process.env.EMAIL_FROM || "SISI <noreply@sisiciechanow.pl>";
 
 type Mail = {
-  from: string;
+  from?: string;
   to: string | string[];
   subject: string;
   html?: string;
@@ -23,12 +23,12 @@ export function getTransport() {
       async sendMail(msg: Mail) {
         console.log("[mailer] using Resend", { to: msg.to, subject: msg.subject });
         const r = await resend.emails.send({
-          from: msg.from,
+          from: msg.from || EMAIL_FROM,
           to: Array.isArray(msg.to) ? msg.to : [msg.to],
           subject: msg.subject,
           html: msg.html,
           text: msg.text,
-        });
+        } as any);
         if ((r as any)?.error) throw new Error((r as any).error.message || "Resend error");
         return r;
       },
@@ -45,10 +45,16 @@ export function getTransport() {
     return {
       async sendMail(msg: Mail) {
         console.log("[mailer] using SMTP", { to: msg.to, subject: msg.subject });
-        return transporter.sendMail(msg);
+        return transporter.sendMail({ from: msg.from || EMAIL_FROM, ...msg });
       },
     };
   }
 
   throw new Error("No mail transport configured (set RESEND_API_KEY or SMTP_* envs).");
+}
+
+/** Helper do prostych wysyłek (używa getTransport) */
+export async function sendEmail(opts: Mail) {
+  const tr = getTransport();
+  return tr.sendMail({ from: EMAIL_FROM, ...opts });
 }
